Refactor to handle additional data from USPS.
USPS API returns a text message for errors, but also for other
conditions.  These XML elements do not seem to be documented in the
USPS documentation, but have been present in this API for years.

This is a backwards compatibility (BC) break, as we now return a
Result object with encompasses the previously returned Address
object and the 2 new USPS text message fields.